feat(tests): move models from Features to Utils
