--- tests/Feature/Models/Author.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace ForestAdmin\LaravelForestAdmin\Tests\Feature\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-/**
- * Class Author
- *
- * @package Laravel-forestadmin
- * @license GNU https://www.gnu.org/licenses/licenses.html
- * @link    https://github.com/ForestAdmin/laravel-forestadmin
- */
-class Author extends Model
-{
-}
