+++ tests/Utils/Models/Image.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace ForestAdmin\LaravelForestAdmin\Tests\Utils\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\MorphTo;
+
+/**
+ * Class Image
+ *
+ * @package Laravel-forestadmin
+ * @license GNU https://www.gnu.org/licenses/licenses.html
+ * @link    https://github.com/ForestAdmin/laravel-forestadmin
+ */
+class Image extends Model
+{
+    protected $fillable = [
+        'name',
+        'url',
+        'imageable_type',
+        'imageable_id',
+    ];
+
+    /**
+     * @return MorphTo
+     */
+    public function imageable(): MorphTo
+    {
+        return $this->morphTo();
+    }
+}
